add routing and style navigation component
