+++ src/routes/navigation/navigation.component.jsx
@@ -1,0 +1,24 @@
+import { Outlet, Link } from "react-router-dom";
+import { ReactComponent as CrwnLogo } from "../../assets/083 crown.svg";
+import "./navigation.styles.scss";
+const Navigation = () => {
+  return (
+    <>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          <div className="logo">
+            <CrwnLogo />
+          </div>
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">
+            Shop
+          </Link>
+        </div>
+      </div>
+      <Outlet />
+    </>
+  );
+};
+
+export default Navigation;
